src/models/UniTS/requirements.txt
einops==0.4.0
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.3
patool==1.12
reformer-pytorch==1.4.4
scikit-learn==1.2.2
scipy==1.10.1
sktime==0.16.1
sympy==1.11.1
tqdm==4.64.1
pyyaml
gluonts[torch]
wandb
timm
torcheval